feat: changes double to BigDecimal

+++ src/gr/aueb/cf/cf9/ch18/bankapp/dto/AccountInsertDTO.java
@@ -1,6 +1,8 @@
 package gr.aueb.cf.cf9.ch18.bankapp.dto;
 
-public record AccountInsertDTO (String iban, double balance) {
+import java.math.BigDecimal;
+
+public record AccountInsertDTO (String iban, BigDecimal balance) {
 
 //    private String iban;
 //    private double balance;
